feat: PackageResponse Integer -> Long으로 수정

+++ src/main/java/com/wnis/linkyway/dto/PackageResponse.java
@@ -14,5 +14,5 @@ public class PackageResponse {
     private String nickname;
     private Long tagId;
     private String tagName;
-    private Integer numberOfCard;
+    private Long numberOfCard;
 }
